[rs-order] fix compilation + relative path issue

+++ rs-order/order/order-domain/src/main/java/fr/cnes/regards/modules/order/domain/dto/DatasetTaskDto.java
@@ -66,11 +66,11 @@ public class DatasetTaskDto {
         this.objectsCount = objectsCount;
     }
 
-    public int getFilesCount() {
+    public long getFilesCount() {
         return filesCount;
     }
 
-    public void setFilesCount(int filesCount) {
+    public void setFilesCount(long filesCount) {
         this.filesCount = filesCount;
     }
